Switches to ruby-mode all Ruby-related configs for performance issues
Having run with enh-ruby-mode for a while, I don't seem to find good defaults performance wise and make it work smoothly, instead it does plenty of calls to 'erm-wait-for-parse'. For this reason, I decided to switch over to 'ruby-mode' which offers me the same functionality I use on a daily basis.