Use lsp-treemacs for the error list
With lsp-treemacs a simple list of errors in the workspace can be
gathered by issuing 'lsp-treemacs-errors-list. Then, making a new
frame with 'make-frame-command and relocating the new frame one can
open the *LSP Error List* buffer in the new frame and continue work on
the main frame while keeping an eye on the error list (updated automatically).